Nbody OpenCL mode uses HAT buffer (We need solution for OpenCL4)
